+++ src/disassemble.c
@@ -37,6 +37,7 @@ static const char *file_path_function_name(uint8_t id) {
         case DIAMOND_FILE_PATH_PUBLISH: return "publish";
         case DIAMOND_FILE_PATH_SYNC: return "sync";
         case DIAMOND_FILE_PATH_RENAME: return "rename";
+        case DIAMOND_FILE_PATH_EXIST: return "exist?";
         default: return "<invalid file path function>";
     }
 }